pyproject.toml
[tool.poetry]
authors = ["Alexander Sidorov <alexander@sidorov.dev>"]
description = "Example: Django + MSSQL + Docker"
license = "MIT"
name = "example_django_mssql_docker"
packages = [{include = "example_django_mssql_docker"}]
version = "2023.06.01"


[tool.poetry.dependencies]
devtools = {extras = ["pygments"], version = "^0.11.0"}
dj-database-url = "^2.0.0"
django = "^4.2.1"
django-mssql-backend = "^2.8.1"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
pydantic = {extras = ["dotenv"], version = "^1.10.7"}
pyodbc = "^4.0.39"
python = "3.11.3"
requests = "^2.30.0"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
psycopg2-binary = "^2.9.6"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.1.1"
black = "^23.3.0"
django-stubs = "^4.2.0"
djangorestframework-stubs = "^3.14.0"
flake8 = "^6.0.0"
flake8-annotations-complexity = "^0.0.7"
flake8-bandit = "^4.1.1"
flake8-black = "^0.3.6"
flake8-bugbear = "^23.3.23"
flake8-builtins = "^2.1.0"
flake8-cognitive-complexity = "^0.1.0"
flake8-comprehensions = "^3.12.0"
flake8-debugger = "^4.1.2"
flake8-executable = "^2.1.3"
flake8-expression-complexity = "^0.0.11"
flake8-fixme = "^1.1.1"
flake8-isort = "^6.0.0"
flake8-multiline-containers = "^0.0.19"
flake8-mutable = "^1.2.0"
flake8-pep3101 = "^2.0.0"
flake8-print = "^5.0.0"
flake8-printf-formatting = "^1.1.2"
flake8-simplify = "^0.20.0"
flake8-variables-names = "^0.0.5"
httpie = "^3.2.1"
ipython = "^8.12.0"
isort = "^5.12.0"
mypy = "^1.3.0"
pep8-naming = "^0.13.3"
pytest = "^7.3.1"
types-psycopg2 = "^2.9.21.10"


[tool.black]
line-length = 79
target-version = ["py310"]
include = '^.+\.pyi?$'
extend-exclude = '''
(
    .*\.local.*
    | .*\.idea.*
    | .*\.vscode.*
)
'''


[tool.isort]
force_single_line = true
line_length = 79
multi_line_output = 3
profile = "black"
skip = [
    ".local",
]


[tool.pytest.ini_options]
cache_dir = ".local/pytest/"
console_output_style = "count"
filterwarnings = [
    "ignore::DeprecationWarning",
]
testpaths = [
    "tests",
]


[tool.mypy]
cache_dir = ".local/mypy/"
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
exclude=[".local/*",]
files = ["."]
no_implicit_optional = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
    "pydantic.mypy",
]
pretty = true
python_version = "3.10"
show_error_context = true
sqlite_cache = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
    "pyodbc.*",
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.django-stubs]
django_settings_module = "project.settings"


[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
